requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
opencv-python
pytube
openai-whisper